requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
distro==1.9.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
openai==1.72.0
pydantic==2.11.3
pydantic_core==2.33.1
python-dotenv==1.1.0
python-telegram-bot==22.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
Telethon==1.38.1
